+++ packages/gov-ui-kit/src/core/components/link/link/link.test.tsx
@@ -1,4 +1,5 @@
-import { fireEvent, render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
+import { userEvent } from '@testing-library/user-event';
 import { Link, type ILinkProps } from '.';
 
 describe('<Link /> component', () => {
@@ -15,13 +16,14 @@ describe('<Link /> component', () => {
         expect(linkElement).toHaveAttribute('href', 'http://example.com');
     });
 
-    it('applies correct classes based on disabled prop', () => {
+    it('applies correct classes based on disabled prop', async () => {
+        const user = userEvent.setup();
         const handleClick = jest.fn();
         render(createTestComponent({ children: 'TEST', disabled: true, onClick: handleClick }));
 
         // eslint-disable-next-line testing-library/no-node-access
         const linkElement = screen.getByText('TEST').closest('a')!;
-        fireEvent.click(linkElement);
+        await user.click(linkElement);
 
         expect(linkElement).toHaveAttribute('aria-disabled', 'true');
         expect(linkElement).not.toHaveAttribute('href');