+++ packages/gov-ui-kit/src/core/components/radioGroup/radioCard/radioCard.test.tsx
@@ -1,4 +1,5 @@
-import { fireEvent, render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
+import { userEvent } from '@testing-library/user-event';
 import { RadioGroup } from '..';
 import { IconType } from '../../icon';
 import { RadioCard, type IRadioCardProps } from './radioCard';
@@ -45,12 +46,13 @@ describe('<RadioCard/> component', () => {
         expect(screen.getByRole('radio')).not.toBeChecked();
     });
 
-    it('renders the RADIO_SELECTED icon when checked', () => {
+    it('renders the RADIO_SELECTED icon when checked', async () => {
+        const user = userEvent.setup();
         render(createTestComponent());
 
         const radioButton = screen.getByRole('radio');
 
-        fireEvent.click(radioButton);
+        await user.click(radioButton);
         const checkedIcon = screen.getByTestId(IconType.RADIO_SELECTED);
 
         expect(checkedIcon).toBeVisible();